Classes implementation finished, this has all the classes and their backend implementation except for class MenuItem is yet to be finished
These classes have comments of "In a real application, this would..." all over them. That is the only part which is not yet done for the application.
These comments must be replaced with either a front end implementation display or with dummy data initialization at the back end.
Or in some cases it would require real life actions such as quality checks that should be performed by a staff/waiter then return the result.
Otherwise, these classes are implemented with their needed methods.